requirements.txt
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.48.0
accelerate>=1.2.0
peft>=0.14.0
diffusers>=0.32.0
datasets>=3.2.0
Pillow>=10.0.0
numpy>=1.26.0
scipy>=1.14.0
tqdm>=4.66.0
wandb>=0.19.0
einops>=0.8.0
timm>=1.0.0
pycocotools>=2.0.8
pyyaml>=6.0
